move the res.socket.end() drain tests to their own file
node-http.test.ts has a pre-existing localhost-resolution failure in the
'request via http proxy, issue#4295' test in this environment, which
trips the fail-with-fix check even though it is unrelated to this
change. A dedicated file gives a clean 5 fail / 0 fail proof on both the
ASAN and release builds.
